+++ clash-protocols/src/Protocols/Experimental/Wishbone/Standard/Hedgehog.hs
@@ -30,7 +30,7 @@ This "common sense" compliance checking additionally checks for:
    - response data respects the 'busSelect' signal
  - A write request must contain valid data according to the 'busSelect' signal
 -}
-module Protocols.Wishbone.Standard.Hedgehog (
+module Protocols.Experimental.Wishbone.Standard.Hedgehog (
   -- * Types
   WishboneMasterRequest (..),
 
@@ -71,8 +71,8 @@ import Hedgehog qualified as H
 import Hedgehog.Gen qualified as Gen
 import Hedgehog.Range qualified as Range
 import Protocols hiding (circuit, stallC)
+import Protocols.Experimental.Wishbone
 import Protocols.Hedgehog
-import Protocols.Wishbone
 import Prelude as P hiding (cycle)
 
 -- | Datatype representing a single transaction request sent from a Wishbone Master to a Wishbone Slave
@@ -99,8 +99,11 @@ deriving instance
   (Eq (WishboneMasterRequest addressBits dataBytes))
 
 {- | Checks equality for relevant parts of a t'WishboneS2M' response based on the
-corresponding 'WishboneMasterRequest'. For a 'Protocols.Wishbone.Standard.Hedgehog.Write' request, the 'readData' field
-is ignored, for a 'Protocols.Wishbone.Standard.Hedgehog.Read' request only the selected bytes are checked.
+corresponding 'WishboneMasterRequest'. For a
+'Protocols.Experimental.Wishbone.Standard.Hedgehog.Write' request, the
+'readData' field is ignored, for a
+'Protocols.Experimental.Wishbone.Standard.Hedgehog.Read' request only the
+selected bytes are checked.
 
 >>>
 :{
